AP-7811: Further implementation of dialog and node
Squashed commits:
Generalize zoom, move resize button, and update width/height
[a7d0f89e] AP-7811: Add radio buttons for colors for clusters or table
[d4bfbd5e] AP-7811: Switch orientation to enum
[9360dd48] AP-7811: Add option to subscribe to filter events
[ac7a5621] AP-7811: Add color palette option
[d2206698] AP-7811: Add a utility class for working with color palettes
[0ddedcc6] AP-7811: Re-organize dialog components
[9ede70ff] AP-7811: Support additional orientations (horizontal, etc.)
[c688030a] AP-7811: Add support for normalized distances
[5fa42837] AP-7811: Fix execute to not save view value to config
[43c4b8c5] AP-7811: Switch scale options to use log scale

There are only two options for y-axis scale: linear and log.
[fc01bdae] AP-7811: Display number mismatch rows/leaves in warning
[29e61e91] AP-7811: Handle more table rows than tree leaves and vice versa

In either case, the given cluster tree model can be used to create a
dendrogram for the view. When the tree has leaves that don't correspond
to rows in the table, this is noted in the JSON and a warning is given.
When the table has rows that don't correspond to leaves in the tree, a
warning is displayed and those rows have missing values for cluster
assignment and selection columns.

AP-7811: Javascript Dendrogram View