AP-10181: fixes in node and dialog
Squashed commits:
Switch to not return 0 for threshold
[bb2e4de3] AP-10181: Switch from Visualizer to Predictor
[716d4062] AP-10181: Fix bug with dialog changes

Previously, when opening the dialog after running the view and saving
the view settings as default the node would detect changes in the
dialog. Even if nothing was changed, and the node would then switch from
an executed state to a configured state.
[813344c8] AP-10181: Fix spacing in dialog
[bd65a3ce] AP-10181: Move all params not in dialog to view value
[20f6deec] AP-10181: Modify cluster assigner to use LayoutTemplateProvider
[6516052e] AP-10181: Add data table with filter columns to JSON
[08650236] AP-10181: Add running in view parameter
[2878515e] AP-10181: Add filter ids and re-build tree on reload
[5cf89917] AP-10181: Switch to PortObjectHolder
[c6e724f4] AP-10181: Switch equals check to equalStructure()

This allows the node to work with tables which contains colors, filters,
     etc. that weren't present in the original data table used to define
     the clusters. This is useful for modifying filters in metanodes,
     without having to re-cluster the data every time.

AP-10181 (JavaScript Hierarchical Cluster Assigner improvements)