AP-10730: Deactivate one-time measurement of labels for columns in meta nodes
In meta nodes, the inital height of the svg is not the final height, as the iframe resizer kicks in.
Therefore we currently can't use the initial measurement of column labels as the final measurement.
See also AP-10592 for the initial reason to use a fixed measurement.

AP-10730 (HeatMap Legend Discrete Values Overlap)